[FIX]update plants and [ADD] dataLoader

+++ src/main/java/com/techforb/plantsMonitor/plant/dto/PlantResponse.java
@@ -16,5 +16,5 @@ public class PlantResponse {
     private Long totalReadings;
     private Integer totalMediumAlerts;
     private Integer totalRedAlerts;
-    private Byte sensorsDisabled;
+    private Integer sensorsDisabled;
 }
